+++ src/ext/java/org/opentripplanner/ext/legacygraphqlapi/LegacyGraphQLIndex.java
@@ -49,7 +49,7 @@ class LegacyGraphQLIndex {
       .setNameFormat("GraphQLExecutor-%d")
       .build());
 
-  static private GraphQLSchema buildSchema() {
+  static protected GraphQLSchema buildSchema() {
     try {
       URL url = Resources.getResource("legacygraphqlapi/schema.graphqls");
       String sdl = Resources.toString(url, Charsets.UTF_8);
@@ -57,6 +57,7 @@ class LegacyGraphQLIndex {
       RuntimeWiring runtimeWiring = RuntimeWiring
           .newRuntimeWiring()
           .scalar(LegacyGraphQLScalars.polylineScalar)
+          .scalar(LegacyGraphQLScalars.geoJsonScalar)
           .scalar(LegacyGraphQLScalars.graphQLIDScalar)
           .scalar(ExtendedScalars.GraphQLLong)
           .type("Node", type -> type.typeResolver(new LegacyGraphQLNodeTypeResolver()))
@@ -101,6 +102,7 @@ class LegacyGraphQLIndex {
           .type("AlertEntity", type -> type.typeResolver(new LegacyGraphQLAlertEntityTypeResolver()))
           .type(IntrospectionTypeWiring.build(LegacyGraphQLStopOnRouteImpl.class))
           .type(IntrospectionTypeWiring.build(LegacyGraphQLStopOnTripImpl.class))
+          .type(IntrospectionTypeWiring.build(LegacyGraphQLStopGeometriesImpl.class))
           .build();
       SchemaGenerator schemaGenerator = new SchemaGenerator();
       return schemaGenerator.makeExecutableSchema(typeRegistry, runtimeWiring);