+++ src/main/java/org/opentripplanner/gtfs/RepairStopTimesForEachTripOperation.java
@@ -2,7 +2,11 @@ package org.opentripplanner.gtfs;
 
 import gnu.trove.list.TIntList;
 import gnu.trove.list.array.TIntArrayList;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 import org.opentripplanner.common.geometry.SphericalDistanceLibrary;
+import org.opentripplanner.ext.flex.trip.FlexTrip;
 import org.opentripplanner.graph_builder.DataImportIssue;
 import org.opentripplanner.graph_builder.DataImportIssueStore;
 import org.opentripplanner.graph_builder.issues.HopSpeedFast;
@@ -15,16 +19,13 @@ import org.opentripplanner.model.Stop;
 import org.opentripplanner.model.StopTime;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.TripStopTimes;
+import org.opentripplanner.util.OTPFeature;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-
 /**
  * This class is responsible for cleaning stop times, removing duplicates, correcting bad data
- * and so on. This do only apply to GTFS imports.
+ * and so on. This only applies to GTFS imports.
  */
 public class RepairStopTimesForEachTripOperation {
     private static final Logger LOG = LoggerFactory.getLogger(RepairStopTimesForEachTripOperation.class);
@@ -52,8 +53,11 @@ public class RepairStopTimesForEachTripOperation {
             /* Fetch the stop times for this trip. Copy the list since it's immutable. */
             List<StopTime> stopTimes = new ArrayList<>(stopTimesByTrip.get(trip));
 
-            /* Remove stoptimes without stops */
-            stopTimes.removeIf(st -> !(st.getStop() instanceof Stop));
+            // if we don't have flex routing enabled then remove all the flex locations and location
+            // groups
+            if(OTPFeature.FlexRouting.isOff()) {
+                stopTimes.removeIf(st -> !(st.getStop() instanceof Stop));
+            }
 
             /* Stop times frequently contain duplicate, missing, or incorrect entries. Repair them. */
             TIntList removedStopSequences = removeRepeatedStops(stopTimes);
@@ -61,6 +65,7 @@ public class RepairStopTimesForEachTripOperation {
                 issueStore.add(new RepeatedStops(trip, removedStopSequences));
             }
             filterStopTimes(stopTimes);
+
             interpolateStopTimes(stopTimes);
 
             stopTimesByTrip.replace(trip, stopTimes);
@@ -258,7 +263,7 @@ public class RepairStopTimesForEachTripOperation {
 
             /* Interpolate, if necessary, the times of non-timepoint stops */
             /* genuine interpolation needed */
-            if (!(st0.isDepartureTimeSet() && st0.isArrivalTimeSet())) {
+            if (!(st0.isDepartureTimeSet() && st0.isArrivalTimeSet()) && !FlexTrip.isFlexStop(st0.getStop())) {
                 // figure out how many such stops there are in a row.
                 int j;
                 StopTime st = null;